Update cybhub.css
Makes the theme adaptive, now it supports a dark and a light variant according to user preference. 
Changes green text to green.
Removes zombie variables.
Improves the primary button and navbar readability, now the text is bold and is no longer italic.
Replaces the drop shadow in floating elements with the highlight border to improve consistency.